refactoring watchers and the start of autoencoders
