+++ app/helpers/utils.py
@@ -9,25 +9,6 @@ import os
 import validators
 
 
-class FileModificationWatcher:
-    _previous_mtimes = {}
-
-    def __init__(self, files=[]):
-        self.add_files(files)
-
-    def add_files(self, files):
-        for f in files:
-            self._previous_mtimes[f] = os.path.getmtime(f)
-
-    def files_changed(self):
-        changed = []
-        for f in self._previous_mtimes:
-            if self._previous_mtimes[f] != os.path.getmtime(f):
-                self._previous_mtimes[f] = os.path.getmtime(f)
-                changed.append(self._previous_mtimes[f])
-        return changed
-
-
 def flatten_dict(d, parent_key='', sep='.'):
     items = []
     for k, v in d.items():
