+++ defaults/outliers.conf
@@ -4,7 +4,7 @@
 [general]
 
 # Elasticsearch parameters
-es_url=http://escoordinator:9200
+es_url=http://esnode1:9200
 es_index_pattern=logstash-eagleeye-*
 es_scan_size=10000
 es_scroll_time=25m
@@ -14,27 +14,27 @@ es_timeout=300
 timestamp_field=timestamp
 
 # Save outlier detection results to Elasticsearch (if set to 0, Elasticsearch events won't be touched - great for testing)
-es_save_results=1
+es_save_results=0
 
 # Print outlier matches to the console. For testing purposes, it's advised to enable this so that the analyst can directly see on the command line which outliers are detected
 print_outliers_to_console=1
 
 # How far back should we process events and look for outliers?
 # Both values are combined (for example the below will look back 7 days and 12 hours, up until right now).
-history_window_days=1
+history_window_days=7
 history_window_hours=12
 
 # Wipe all existing outliers that fall in the history window upon first run
-es_wipe_all_existing_outliers=1
+es_wipe_all_existing_outliers=0
 
 # How often should existing outliers be checked for a match against the configuration whitelist, and should outlier fields be removed again from matched documents
-housekeeping_interval_seconds=60
-es_wipe_all_whitelisted_outliers=1
+housekeeping_interval_seconds=3600
+es_wipe_all_whitelisted_outliers=0
 
 # General flags to run, test or train models
 # Training is only used by the Machine Learning models (word2vec)
-run_models=1
-test_models=0
+run_models=0
+test_models=1
 train_models=0
 
 # 0 for no progress info,  1-4 for progressively more output, 5+ for all the log output
@@ -252,6 +252,25 @@ outlier_summary=base64 encoded command line arguments for process {OsqueryFilter
 
 run_model=0
 test_model=0
+
+######################################################################################################################################################
+# AUTOENCODER USE CASE MODEL - EXAMPLES
+######################################################################################################################################################
+
+##############################
+# AUTOENCODER - DETECT SUSPICIOUS HOST NAME
+##############################
+[autoencoder_find_suspicious_host_names]
+es_query_filter= tags:endpoint AND _exists_: meta.hostname
+aggregator= meta.deployment_name
+target = meta.hostname
+
+outlier_type=anomaly
+outlier_reason=suspicious host name
+outlier_summary=suspicious host name {meta.hostname}
+
+run_model = 0
+test_model = 1
 
 ######################################################################################################################################################
 # WHITELISTS
